+++ src/graphql/resolvers/seedling.ts
@@ -1,13 +1,14 @@
-import mongoose from 'mongoose';
 import parseDate from "../../utils/parseDate";
 import Seedling from '../../models/seedling';
-import { Request } from "express";
 
 export default {
-    seedlings: async (args: any, req: Request) => {
+    seedlings: async (args: any, req: any) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthorized!');
+        }
         return Seedling.find().then(seedlings => {
             return seedlings.map(seedling => {
-                return { ...seedling, daysInSoil: parseDate(seedling.datePlanted) }
+                return { ...seedling._doc, daysInSoil: parseDate(seedling.datePlanted) }
             });
         }).catch(error => {
             console.error(error);
@@ -15,21 +16,26 @@ export default {
         });
     },
 
-    oneSeedling: async (args: any, req: Request) => {
+    oneSeedling: async (args: any, req: any) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthorized!');
+        }
         return Seedling.findById(args._id).then(seedling => {
             if (!seedling) {
                 throw new Error('Seedling haven\'t been found');
             }
-            return { ...seedling, daysInSoil: parseDate(seedling.datePlanted) }
+            return { ...seedling._doc, daysInSoil: parseDate(seedling.datePlanted) }
         }).catch(error => {
             console.error(error);
             throw error;
         });
     },
 
-    createSeedling: async (args: any, req: Request) => {
+    createSeedling: async (args: any, req: any) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthorized!');
+        }
         const seedling = new Seedling({
-            _id: mongoose.Types.ObjectId(),
             species: args.seedlingInput.species,
             plantedQuantity: +args.seedlingInput.plantedQuantity,
             survivedQuantity: +args.seedlingInput.plantedQuantity,
@@ -39,14 +45,17 @@ export default {
             pictureId: args.seedlingInput.pictureId
         });
         return seedling.save().then(seedling => {
-            return { ...seedling, daysInSoil: parseDate(seedling.datePlanted) }
+            return { ...seedling._doc, daysInSoil: parseDate(seedling.datePlanted) }
         }).catch(error => {
             console.error(error);
             throw error;
         });
     },
 
-    updateSeedling: async (args: any, req: Request) => {
+    updateSeedling: async (args: any, req: any) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthorized!');
+        }
         const id = args.seedlingInput._id;
         const updateSeedling = {
             species: args.seedlingInput.species,
@@ -62,14 +71,17 @@ export default {
             if (!updatedSeedling) {
                 throw new Error('Seedling doesn\'t exist');
             };
-            return { ...updatedSeedling, daysInSoil: parseDate(updatedSeedling.datePlanted) }
+            return { ...updatedSeedling._doc, daysInSoil: parseDate(updatedSeedling.datePlanted) }
         } catch (error) {
             console.log(error);
             throw error;
         };
     },
 
-    deleteSeedling: async (args: any, req: Request) => {
+    deleteSeedling: async (args: any, req: any) => {
+        if (!req.isAuth) {
+            throw new Error('Unauthorized!');
+        }
         return Seedling.deleteOne({ _id: args._id }).then(deletedSeedling => {
             if (deletedSeedling.deletedCount !== 1) {
                 throw new Error('Delete was unsuccessful');
